@@ -6,7 +6,7 @@
   "scripts": {
     "lint": "eslint src --ext .js --ext .jsx",
     "start": "webpack-dev-server --host 0.0.0.0 --port 8080 --hot --progress",
-    "build": "npm run lint; NODE_ENV=production webpack --optimize-minimize --progress --colors",
+    "build": "npm run lint; NODE_ENV=production webpack --progress --colors",
     "dev-build": "npm run lint; webpack --progress --colors",
     "start-webdriver": "webdriver-manager update && webdriver-manager start",
     "test:only": "ava --verbose --serial",
@@ -20,12 +20,14 @@
     "babel-core": "^6.3.17",
     "babel-eslint": "^7.1.0",
     "babel-loader": "^7.0.0",
-    "babel-plugin-transform-object-rest-spread": "^6.20.2",
-    "babel-preset-es2017-node7": "^0.5.2",
+    "babel-plugin-react-require": "^3.0.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
     "css-loader": "^0.28.7",
     "style-loader": "^0.18.2",
-    "less-loader": "^4.0.5",
+    "less-loader": "^4.0.4",
     "eslint": "^3.8.1",
     "eslint-loader": "^1.1.1",
     "eslint-plugin-react": "^7.0.0",
@@ -35,7 +37,6 @@
     "http-server": "^0.9.0",
     "json-loader": "^0.5.4",
     "less": "^2.7.1",
-    "less-loader": "^4.0.4",
     "node-libs-browser": "^2.0.0",
     "parallelshell": "^2.0.0",
     "react-hot-loader": "^3.0.0-beta.6",
@@ -66,13 +67,14 @@
     "octicons": "^6.0.0",
     "octokat": "^0.7.0",
     "puzzle-script": "0.0.2",
+    "prop-types": "^15.6.1",
     "qs": "^6.3.0",
-    "react": "^15.3.2",
-    "react-bootstrap": "^0.31.0",
+    "react": "^16.3.2",
+    "react-bootstrap": "^0.31.5",
     "react-dnd": "^2.1.2",
     "react-dnd-html5-backend": "^2.1.2",
-    "react-dom": "^15.3.2",
-    "react-octicons": "0.0.5",
+    "react-dom": "^16.3.2",
+    "react-octicons": "0.2.0",
     "react-router": "^3.0.0",
     "recordo": "^0.0.6",
     "ultramarked": "^1.7.0",
