@@ -1,5 +1,5 @@
 import _ from 'underscore';
-import React from 'react';
+import {Component} from 'react';
 import {Link} from 'react-router';
 import * as BS from 'react-bootstrap';
 import {HomeIcon, StarIcon, GearIcon, QuestionIcon, GraphIcon, TagIcon} from 'react-octicons';
@@ -16,44 +16,43 @@ import MoveModal from '../move-modal';
 import FilterDropdown from './filter-dropdown';
 
 
-const SettingsItem = React.createClass({
-  render() {
-    const {key, onSelect, isChecked, className, to, children} = this.props;
-    let {href} = this.props;
+function SettingsItem(props) {
+  const {key, onSelect, isChecked, className, to, children} = props;
+  let {href} = props;
 
-    if (!href && to) {
-      href = `#${to}`; // Link
-    }
-
-    return (
-      <BS.MenuItem key={key} href={href} onSelect={onSelect} className={className}>
-        <span className='settings-item-checkbox' data-checked={isChecked}>{children}</span></BS.MenuItem>
-    );
+  if (!href && to) {
+    href = `#${to}`; // Link
   }
-});
 
+  return (
+    <BS.MenuItem key={key} href={href} onSelect={onSelect} className={className}>
+      <span className='settings-item-checkbox' data-checked={isChecked}>{children}</span></BS.MenuItem>
+  );
+}
 
-const AppNav = React.createClass({
-  getInitialState() {
-    return {info: null, showModal: false};
-  },
+class AppNav extends Component {
+  state = {info: null, showModal: false};
+
   componentDidMount() {
     SettingsStore.on('change', this.update);
     SettingsStore.on('change:showPullRequestData', this.update);
     SettingsStore.on('change:tableLayout', this.update);
     Client.on('changeToken', this.onChangeToken);
     this.onChangeToken();
-  },
+  }
+
   componentWillUnmount() {
     SettingsStore.off('change', this.update);
     SettingsStore.off('change:showPullRequestData', this.update);
     SettingsStore.off('change:tableLayout', this.update);
     Client.off('changeToken', this.onChangeToken);
-  },
-  update() {
+  }
+
+  update = () => {
     this.setState({});
-  },
-  onChangeToken() {
+  };
+
+  onChangeToken = () => {
     CurrentUserStore.fetchUser()
     .then((info) => {
       // TODO: when anonymous, getting the current user should be an error.
@@ -65,25 +64,29 @@ const AppNav = React.createClass({
     }).catch(() => {
       this.setState({info: null});
     });
-  },
-  onSignOut() {
+  };
+
+  onSignOut = () => {
     Client.setToken(null);
     CurrentUserStore.clear();
-  },
-  starThisProject() {
+  };
+
+  starThisProject = () => {
     Client.getOcto().user.starred('philschatz/gh-board').add().then(() => {
       /*eslint-disable no-alert */
       alert('Thanks for starring!\n I hope you enjoy the other pages more than this simple alert, but thank you for helping me out!');
       /*eslint-enable no-alert */
     });
-  },
-  promptAndResetDatabases() {
+  };
+
+  promptAndResetDatabases = () => {
     if (confirm('Are you sure you want to reset all the local data? It will take some time to repopulate all the data from GitHub and you may need to reload the page')) {
       Database.resetDatabases().then(() => {
         alert('Local cache has been cleared');
       });
     }
-  },
+  };
+
   render() {
     let routeInfo = getFilters().getState();
     let {repoInfos} = routeInfo;
@@ -281,7 +284,6 @@ const AppNav = React.createClass({
       </div>
     );
   }
-
-});
+}
 
 export default AppNav;
