@@ -1,18 +1,19 @@
-import React from 'react';
+import {Component} from 'react';
 import * as BS from 'react-bootstrap';
 import {TagIcon} from 'react-octicons';
 
 const LOCALSTORAGE_KEY = 'saved-filters';
 
-const AddFilterModal = React.createClass({
-  onSave() {
+class AddFilterModal extends Component {
+  onSave = () => {
     const {onHide} = this.props;
     const data = JSON.parse(window.localStorage[LOCALSTORAGE_KEY] || '[]');
     const title = this._title.value;
     data.push({title: title, hashStr: window.location.hash});
     window.localStorage[LOCALSTORAGE_KEY] = JSON.stringify(data);
     onHide();
-  },
+  };
+
   render() {
     return (
       <BS.Modal className='-add-filter-modal' {...this.props}>
@@ -28,18 +29,19 @@ const AddFilterModal = React.createClass({
       </BS.Modal>
     );
   }
-});
+}
 
-const SavedFiltersButton = React.createClass({
-  getInitialState() {
-    return {showModal: false};
-  },
-  showAddFilter() {
+class SavedFiltersButton extends Component {
+  state = {showModal: false};
+
+  showAddFilter = () => {
     this.setState({showModal: true});
-  },
-  hideAddFilter() {
+  };
+
+  hideAddFilter = () => {
     this.setState({showModal: false});
-  },
+  };
+
   render() {
     const {showModal} = this.state;
     const savedFilterData = JSON.parse(window.localStorage['saved-filters'] || '[]');
@@ -62,6 +64,6 @@ const SavedFiltersButton = React.createClass({
       </li>
     );
   }
-});
+}
 
 export default SavedFiltersButton;
