@@ -1,4 +1,4 @@
-import React from 'react';
+import {Component} from 'react';
 import * as BS from 'react-bootstrap';
 import {SyncIcon} from 'react-octicons';
 
@@ -8,35 +8,36 @@ const ASYNC_REJECTED = 'ASYNC_REJECTED';
 const ASYNC_WAITING = 'ASYNC_WAITING';
 
 
-export default React.createClass({
-  getInitialState() {
-    return {val: null, asyncStatus: ASYNC_UNSTARTED};
-  },
-  getDefaultProps() {
-    return {
-      waitingText: 'Saving...'
-    };
-  },
+class AsyncButton extends Component {
+  static defaultProps = {
+    waitingText: 'Saving...'
+  };
+
+  state = {val: null, asyncStatus: ASYNC_UNSTARTED};
+
   componentDidUpdate(oldProps) {
     if (this.props.action !== oldProps.action) {
       this.setState({val: null, asyncStatus: ASYNC_UNSTARTED});
     }
-  },
-  onResolve(val) {
+  }
+
+  onResolve = (val) => {
     const {onResolved} = this.props;
     this.setState({val, asyncStatus: ASYNC_RESOLVED});
     if (onResolved) {
       onResolved(val);
     }
-  },
-  onReject(val) {
+  };
+
+  onReject = (val) => {
     const {onRejected} = this.props;
     this.setState({val, asyncStatus: ASYNC_REJECTED});
     if (onRejected) {
       onRejected(val);
     }
-  },
-  onClick() {
+  };
+
+  onClick = () => {
     const {action} = this.props;
     this.setState({val: null, asyncStatus: ASYNC_WAITING});
     const p = action();
@@ -44,7 +45,8 @@ export default React.createClass({
     if (p) {
       p.then(this.onResolve, this.onReject);
     }
-  },
+  };
+
   render() {
     const {children, waitingText, renderError, doneText, className, ...rest} = this.props;
     let {disabled} = this.props;
@@ -87,4 +89,6 @@ export default React.createClass({
         >{kids}</BS.Button>
     );
   }
-});
+}
+
+export default AsyncButton;
