@@ -1,4 +1,4 @@
-import React from 'react';
+import {Component} from 'react';
 import _ from 'underscore';
 import {Link} from 'react-router';
 import * as BS from 'react-bootstrap';
@@ -19,13 +19,13 @@ const SAMPLE_REPOS = [
 ];
 
 
-const ListGroupWithMore = React.createClass({
-  getInitialState() {
-    return {morePressedCount: 0};
-  },
-  onClickMore() {
+class ListGroupWithMore extends Component {
+  state = {morePressedCount: 0};
+
+  onClickMore = () => {
     this.setState({morePressedCount: this.state.morePressedCount + 1});
-  },
+  };
+
   render() {
     const {children} = this.props;
     const {morePressedCount} = this.state;
@@ -49,76 +49,73 @@ const ListGroupWithMore = React.createClass({
       </BS.ListGroup>
     );
   }
-});
+}
 
-const RepoItem = React.createClass({
-  render() {
-    const {repoOwner, comment, isSelected, onSelect} = this.props;
-    let {repoName, repoNames} = this.props;
+function RepoItem(props) {
+  const {repoOwner, comment, isSelected, onSelect} = props;
+  let {repoName, repoNames} = props;
 
-    // Example repos do not have additional information and may contain multiple repos
-    let {repo} = this.props;
-    repo = repo || {};
-    if (repoNames) {
-      repoName = repoNames.join(' & ');
-    } else {
-      repoNames = [repoName];
-    }
-    const repoInfos = repoNames.map((name) => {
-      return {repoOwner, repoName: name};
-    });
+  // Example repos do not have additional information and may contain multiple repos
+  let {repo} = props;
+  repo = repo || {};
+  if (repoNames) {
+    repoName = repoNames.join(' & ');
+  } else {
+    repoNames = [repoName];
+  }
+  const repoInfos = repoNames.map((name) => {
+    return {repoOwner, repoName: name};
+  });
 
 
-    let repoIcon;
-    if (repo.private) { repoIcon = (<LockIcon className='repo-icon'/>); }
-    else if (repo.fork) { repoIcon = (<RepoForkedIcon className='repo-icon'/>); }
-    else { repoIcon = (<RepoIcon className='repo-icon'/>); }
+  let repoIcon;
+  if (repo.private) { repoIcon = (<LockIcon className='repo-icon'/>); }
+  else if (repo.fork) { repoIcon = (<RepoForkedIcon className='repo-icon'/>); }
+  else { repoIcon = (<RepoIcon className='repo-icon'/>); }
 
-    const classes = {
-      'repo-item': true,
-      'is-private': repo.private
-    };
+  const classes = {
+    'repo-item': true,
+    'is-private': repo.private
+  };
 
-    let updatedAt = null;
-    if (repo.pushedAt) {
-      updatedAt = (
-        <small className='repo-updated-at'>
-          {'updated '}
-          <Time dateTime={repo.pushedAt}/>
-        </small>
-      );
-    }
-
-    let multiSelectButton = null;
-    if (onSelect) {
-      multiSelectButton = (
-        <input
-          className='pull-right'
-          type='checkbox'
-          checked={isSelected}
-          onChange={onSelect}/>
-      );
-    }
-
-    const repoLink = buildRoute('kanban', {repoInfos});
-    return (
-      <BS.ListGroupItem key={repoName} className={classnames(classes)}>
-        {repoIcon}
-        <Link to={repoLink}>{repoName}</Link>
-        {repo.private && (<BS.Label className='repo-private-label' bsStyle='warning'>PRIVATE</BS.Label>) || null} {' '}
-        {updatedAt}
-        {comment}
-        {multiSelectButton}
-      </BS.ListGroupItem>
+  let updatedAt = null;
+  if (repo.pushedAt) {
+    updatedAt = (
+      <small className='repo-updated-at'>
+        {'updated '}
+        <Time dateTime={repo.pushedAt}/>
+      </small>
     );
   }
-});
 
-const RepoGroup = React.createClass({
-  getInitialState() {
-    return {selectedRepos: {}};
-  },
-  toggleSelect(repoName) {
+  let multiSelectButton = null;
+  if (onSelect) {
+    multiSelectButton = (
+      <input
+        className='pull-right'
+        type='checkbox'
+        checked={isSelected}
+        onChange={onSelect}/>
+    );
+  }
+
+  const repoLink = buildRoute('kanban', {repoInfos});
+  return (
+    <BS.ListGroupItem key={repoName} className={classnames(classes)}>
+      {repoIcon}
+      <Link to={repoLink}>{repoName}</Link>
+      {repo.private && (<BS.Label className='repo-private-label' bsStyle='warning'>PRIVATE</BS.Label>) || null} {' '}
+      {updatedAt}
+      {comment}
+      {multiSelectButton}
+    </BS.ListGroupItem>
+  );
+}
+
+class RepoGroup extends Component {
+  state = {selectedRepos: {}};
+
+  toggleSelect = (repoName) => {
     return () => {
       const {selectedRepos} = this.state;
       if (selectedRepos[repoName]) {
@@ -129,15 +126,17 @@ const RepoGroup = React.createClass({
       this.setState({selectedRepos});
       // this.forceUpdate(); // since we are modifying hte object directly
     };
-  },
-  goToBoard() {
+  };
+
+  goToBoard = () => {
     const {repoOwner} = this.props;
     const {selectedRepos} = this.state;
     const repoInfos = Object.keys(selectedRepos).map((repoName) => {
       return {repoOwner, repoName};
     });
     history.push(buildRoute('kanban', {repoInfos}));
-  },
+  };
+
   render() {
     let {repoOwner, repos, index} = this.props;
     const {selectedRepos} = this.state;
@@ -193,68 +192,66 @@ const RepoGroup = React.createClass({
       </BS.Col>
     );
   }
-});
+}
 
-const Dashboard = React.createClass({
-  displayName: 'Dashboard',
-  render() {
-    const {repos} = this.props;
+function Dashboard(props) {
+  const {repos} = props;
 
-    const repoOwnersUpdatedAt = {};
-    const reposByOwner = {};
+  const repoOwnersUpdatedAt = {};
+  const reposByOwner = {};
 
-    _.each(repos, (repo) => {
-      const repoOwner = repo.owner.login;
-      const updatedAt = repo.pushedAt;
+  _.each(repos, (repo) => {
+    const repoOwner = repo.owner.login;
+    const updatedAt = repo.pushedAt;
 
-      if (!repoOwnersUpdatedAt[repoOwner] || repoOwnersUpdatedAt[repoOwner] < updatedAt) {
-        repoOwnersUpdatedAt[repoOwner] = updatedAt;
-      }
-      if (!reposByOwner[repoOwner]) {
-        reposByOwner[repoOwner] = [];
-      }
-      reposByOwner[repoOwner].push(repo);
-    });
+    if (!repoOwnersUpdatedAt[repoOwner] || repoOwnersUpdatedAt[repoOwner] < updatedAt) {
+      repoOwnersUpdatedAt[repoOwner] = updatedAt;
+    }
+    if (!reposByOwner[repoOwner]) {
+      reposByOwner[repoOwner] = [];
+    }
+    reposByOwner[repoOwner].push(repo);
+  });
 
-    let sortedRepoOwners = _.map(repoOwnersUpdatedAt, (updatedAt, repoOwner) => {
-      return {repoOwner, updatedAt};
-    });
-    sortedRepoOwners = _.sortBy(sortedRepoOwners, ({updatedAt}) => {
-      return updatedAt;
-    });
-    sortedRepoOwners.reverse();
+  let sortedRepoOwners = _.map(repoOwnersUpdatedAt, (updatedAt, repoOwner) => {
+    return {repoOwner, updatedAt};
+  });
+  sortedRepoOwners = _.sortBy(sortedRepoOwners, ({updatedAt}) => {
+    return updatedAt;
+  });
+  sortedRepoOwners.reverse();
 
 
-    const repoOwnersNodes = _.map(sortedRepoOwners, ({repoOwner /*,updatedAt*/}, index) => {
-      const groupRepos = reposByOwner[repoOwner];
-      return (
-        <RepoGroup
-          key={repoOwner}
-          repoOwner={repoOwner}
-          repos={groupRepos}
-          index={index}/>
-      );
-    });
-
+  const repoOwnersNodes = _.map(sortedRepoOwners, ({repoOwner /*,updatedAt*/}, index) => {
+    const groupRepos = reposByOwner[repoOwner];
     return (
-          <span>{repoOwnersNodes}</span>
+      <RepoGroup
+        key={repoOwner}
+        repoOwner={repoOwner}
+        repos={groupRepos}
+        index={index}/>
     );
-  }
-});
+  });
 
-const CustomRepoModal = React.createClass({
-  getInitialState() {
-    return {customRepoName: null};
-  },
-  onCustomRepoChange(e) {
+  return (
+        <span>{repoOwnersNodes}</span>
+  );
+}
+
+class CustomRepoModal extends Component {
+  state = {customRepoName: null};
+
+  onCustomRepoChange = (e) => {
     this.setState({customRepoName: e.currentTarget.value});
-  },
-  goToBoard(customRepoName) {
+  };
+
+  goToBoard = (customRepoName) => {
     const [repoOwner, repoName] = customRepoName.split('/');
     const repoInfos = [{repoOwner, repoName}];
     // TODO: Just make this a simple Link and no fancy history.push
     history.push(buildRoute('kanban', {repoInfos}));
-  },
+  };
+
   render() {
     const {customRepoName} = this.state;
     // Make sure the repo contains a '/'
@@ -288,15 +285,15 @@ const CustomRepoModal = React.createClass({
       </BS.Modal>
     );
   }
-});
+}
 
-const ExamplesPanel = React.createClass({
-  getInitialState() {
-    return {showModal: false};
-  },
-  onClickMore() {
+class ExamplesPanel extends Component {
+  state = {showModal: false};
+
+  onClickMore = () => {
     this.setState({showModal: true});
-  },
+  };
+
   render() {
     const {showModal} = this.state;
     const close = () => this.setState({ showModal: false});
@@ -321,15 +318,14 @@ const ExamplesPanel = React.createClass({
       </BS.Panel>
     );
   }
-});
+}
 
 
 let allMyReposHack = null;
 
-const DashboardShell = React.createClass({
-  getInitialState() {
-    return {repos: null};
-  },
+class DashboardShell extends Component {
+  state = {repos: null};
+
   render() {
     let {repos} = this.state;
 
@@ -375,6 +371,6 @@ const DashboardShell = React.createClass({
       </BS.Grid>
     );
   }
-});
+}
 
 export default DashboardShell;
