@@ -1,4 +1,4 @@
-import React from 'react';
+import {Component} from 'react';
 import _ from 'underscore';
 import {Link} from 'react-router';
 import {ListUnorderedIcon} from 'react-octicons';
@@ -32,58 +32,57 @@ const filterKanbanLabels = (labels, columnRegExp) => {
 };
 
 
-const KanbanColumn = React.createClass({
-  render() {
-    const {label, cards, primaryRepo} = this.props;
-    const {columnRegExp} = getFilters().getState();
+function KanbanColumn(props) {
+  const {label, cards, primaryRepo} = props;
+  const {columnRegExp} = getFilters().getState();
 
-    const issueComponents = _.map(cards, (card) => {
-      return (
-        <Issue
-          key={card.issue.id}
-          primaryRepoName={primaryRepo.repoName}
-          card={card}
-          columnRegExp={columnRegExp}
-          />
-      );
-    });
-
-    let icon;
-    let name;
-    if (columnRegExp.test(label.name)) {
-      icon = (<ListUnorderedIcon/>);
-      name = label.name.replace(/^\d+\ -\ /, ' ');
-    } else {
-      icon = null;
-      name = label.name;
-    }
-    const title = (
-      <Link className='label-title' to={getFilters().toggleColumnLabel(label.name).url()}>
-        {name}
-      </Link>
+  const issueComponents = _.map(cards, (card) => {
+    return (
+      <Issue
+        key={card.issue.id}
+        primaryRepoName={primaryRepo.repoName}
+        card={card}
+        columnRegExp={columnRegExp}
+        />
     );
+  });
 
-    if (issueComponents.length || SettingsStore.getShowEmptyColumns()) {
-      return (
-        <div key={label.name} className='kanban-board-column'>
-          <IssueList
-            icon={icon}
-            title={title}
-            backgroundColor={label.color}
-            label={label}
-            primaryRepo={primaryRepo}
-          >
-            {issueComponents}
-          </IssueList>
-        </div>
-      );
-    } else {
-      return null; // TODO: Maybe the panel should say "No Issues" (but only if it's the only column)
-    }
+  let icon;
+  let name;
+  if (columnRegExp.test(label.name)) {
+    icon = (<ListUnorderedIcon/>);
+    name = label.name.replace(/^\d+\ -\ /, ' ');
+  } else {
+    icon = null;
+    name = label.name;
   }
-});
+  const title = (
+    <Link className='label-title' to={getFilters().toggleColumnLabel(label.name).url()}>
+      {name}
+    </Link>
+  );
 
-const KanbanRepo = React.createClass({
+
+  if (issueComponents.length || SettingsStore.getShowEmptyColumns()) {
+    return (
+      <div key={label.name} className='kanban-board-column'>
+        <IssueList
+          icon={icon}
+          title={title}
+          backgroundColor={label.color}
+          label={label}
+          primaryRepo={primaryRepo}
+        >
+          {issueComponents}
+        </IssueList>
+      </div>
+    );
+  } else {
+    return null; // TODO: Maybe the panel should say "No Issues" (but only if it's the only column)
+  }
+}
+
+class KanbanRepo extends Component {
   render() {
     const {columnData, cards, repoInfos} = this.props;
     const {columnRegExp} = getFilters().getState();
@@ -137,20 +136,22 @@ const KanbanRepo = React.createClass({
       </div>
     );
   }
-});
+}
 
 
 let showedWarning = false;
 
-const RepoKanbanShell = React.createClass({
-  componentWillMount() {
+class RepoKanbanShell extends Component {
+  UNSAFE_componentWillMount() {
     // Needs to be called before `render()`
     IssueStore.startPolling();
-  },
+  }
+
   componentWillUnmount() {
     IssueStore.stopPolling();
-  },
-  renderLoaded() {
+  }
+
+  renderLoaded = () => {
     const {repoInfos} = getFilters().getState();
     // Get the "Primary" repo for milestones and labels
     const [{repoOwner, repoName}] = repoInfos;
@@ -171,7 +172,8 @@ const RepoKanbanShell = React.createClass({
         columnDataPromise={promise}
       />
     );
-  },
+  };
+
   render() {
 
     return (
@@ -181,6 +183,6 @@ const RepoKanbanShell = React.createClass({
       />
     );
   }
-});
+}
 
 export default RepoKanbanShell;
