@@ -1,24 +1,27 @@
+import c3 from 'c3';
+
 /**
  * React-C3 Chart
  * Copyright 2015 - Cody Reichert <codyreichert@gmail.com>
  */
 
-import c3    from 'c3';
-import React from 'react';
+import PropTypes from 'prop-types';
 
-const ChartComponent = React.createClass({
-  displayName: 'React-C3-Chart',
+import {Component} from 'react';
 
-  propTypes: {
-    className: React.PropTypes.string,
-    data: React.PropTypes.object.isRequired,
-    element: React.PropTypes.string.isRequired,
-    options: React.PropTypes.object
-  },
+class ChartComponent extends Component {
+  static displayName = 'React-C3-Chart';
 
-  chart: null,
+  static propTypes = {
+    className: PropTypes.string,
+    data: PropTypes.object.isRequired,
+    element: PropTypes.string.isRequired,
+    options: PropTypes.object
+  };
 
-  shouldComponentUpdate: function(nextProps) {
+  chart = null;
+
+  shouldComponentUpdate(nextProps) {
     if(this.props.data.rows.length
           !== nextProps.data.rows.length) { // shallow check
       return true;
@@ -27,17 +30,17 @@ const ChartComponent = React.createClass({
       return  true;
     }
     return false;
-  },
+  }
 
-  componentDidMount: function() {
+  componentDidMount() {
     this._generateChart(
       this.props.data,
       this.props.element,
       this.props.options
     );
-  },
+  }
 
-  componentDidUpdate: function(prevProps) {
+  componentDidUpdate(prevProps) {
     if(prevProps.data.rows !== this.props.data.rows) {
       this._generateChart(
         this.props.data,
@@ -45,25 +48,25 @@ const ChartComponent = React.createClass({
         this.props.options
       );
     }
-  },
+  }
 
-  componentWillUnmount: function() {
+  componentWillUnmount() {
     this._destroyChart();
-  },
+  }
 
-  _generateChart: function(data, element, options) {
+  _generateChart = (data, element, options) => {
     let build = Object.assign({}, {
       bindto: '#' + element,
       data: data
     }, options);
     this.chart = c3.generate(build);
-  },
+  };
 
-  _destroyChart: function() {
+  _destroyChart = () => {
     this.chart.destroy();
-  },
+  };
 
-  render: function() {
+  render() {
     let className;
     if (this.props.className) {
       className = 'c3 ' + this.props.className;
@@ -77,7 +80,7 @@ const ChartComponent = React.createClass({
       </div>
     );
   }
-});
+}
 
 
 export default ChartComponent;
