@@ -1,4 +1,4 @@
-import React from 'react';
+import {Component} from 'react';
 import {SyncIcon} from 'react-octicons';
 
 const STATUS = {
@@ -7,63 +7,64 @@ const STATUS = {
   ERROR: 'error'
 };
 
-export default React.createClass({
-  displayName: 'Loadable',
-  getInitialState() {
-    return {status: STATUS.INITIAL, value: null};
-  },
-  getDefaultProps() {
-    return {
-      renderError: (err) => {
-        console.error(err);
-        // If it is a permissions error then it might be a rate limi
-        if (err.status === 403) {
-          return (
-            <div>
-              <h2>Insufficient Permissions (or rate limit exceeded)</h2>
-              <p>
-                It looks like either you do not have permission to see this repository or the rate limit for requests to GitHub has been exceeded. This usually happens when you are not logged in to gh-board. Try signing in to continue.
-              </p>
-              <code>{err.message}</code>
-            </div>
-          );
-        } else if (err.name === 'InvalidStateError') {
-          return (
-            <span>It looks like your browser is in private browsing mode. gh-board uses IndexedDB to cache requests to GitHub. Please disable Private Browsing to see it work.</span>
-          );
-        } else {
-          return (
-            <span>
-              Problem loading. Is it a valid repo? And have you exceeded your number of requests? Usually happens when not logged in because GitHub limits anonymous use of their API.
-              {err.message}
-              {JSON.stringify(err)}
-            </span>
-          );
-        }
+class Loadable extends Component {
+  static defaultProps = {
+    renderError: (err) => {
+      console.error(err);
+      // If it is a permissions error then it might be a rate limi
+      if (err.status === 403) {
+        return (
+          <div>
+            <h2>Insufficient Permissions (or rate limit exceeded)</h2>
+            <p>
+              It looks like either you do not have permission to see this repository or the rate limit for requests to GitHub has been exceeded. This usually happens when you are not logged in to gh-board. Try signing in to continue.
+            </p>
+            <code>{err.message}</code>
+          </div>
+        );
+      } else if (err.name === 'InvalidStateError') {
+        return (
+          <span>It looks like your browser is in private browsing mode. gh-board uses IndexedDB to cache requests to GitHub. Please disable Private Browsing to see it work.</span>
+        );
+      } else {
+        return (
+          <span>
+            Problem loading. Is it a valid repo? And have you exceeded your number of requests? Usually happens when not logged in because GitHub limits anonymous use of their API.
+            {err.message}
+            {JSON.stringify(err)}
+          </span>
+        );
       }
-    };
-  },
+    }
+  };
+
+  state = {status: STATUS.INITIAL, value: null};
+
   componentDidMount() {
     const {promise} = this.props;
     promise.then(this.onResolve, this.onError);
-  },
+  }
+
   componentDidUpdate(prevProps) {
     if (this.props.promise !== prevProps.promise) {
       const {promise} = this.props;
       promise.then(this.onResolve, this.onError);
     }
-  },
-  onResolve(value) {
+  }
+
+  onResolve = (value) => {
     // TODO: Find out why this is being called multiple times
     this.setState({status: STATUS.RESOLVED, value});
-  },
-  onError(value) {
+  };
+
+  onError = (value) => {
     // TODO: Find out why this is being called multiple times
     if (this.state.status !== STATUS.ERROR) {
       this.setState({status: STATUS.ERROR, value});
     }
-  },
-  renderLoading() {
+  };
+
+  renderLoading = () => {
     const {loadingText} = this.props;
     return (
       <span className='loadable is-loading'>
@@ -71,7 +72,8 @@ export default React.createClass({
         {' ' + (loadingText || 'Loading...')}
       </span>
     );
-  },
+  };
+
   render() {
     const {status, value} = this.state;
     let {renderLoading, renderLoaded, renderError} = this.props;
@@ -86,4 +88,6 @@ export default React.createClass({
       return renderError(value);
     }
   }
-});
+}
+
+export default Loadable;
